requirements.txt
discord.py
gspread
oauth2client
apscheduler
python-dotenv
